moving file to exam directories
