+++ jdk11ocp/features/src/main/java/com/vkeonline/enthuware/exam815/ArraysFeatures.java
@@ -1,4 +1,4 @@
-package com.vkeonline.enthuware;
+package com.vkeonline.enthuware.exam815;
 
 public class ArraysFeatures {
     public void checkArrayFeature() {
