+++ jdk11ocp/features/src/main/java/com/vkeonline/enthuware/exam815/InheritanceFeatures.java
@@ -1,4 +1,4 @@
-package com.vkeonline.enthuware;
+package com.vkeonline.enthuware.exam815;
 
 import java.util.ArrayList;
 import java.util.List;
