+++ jdk11ocp/features/src/main/java/com/vkeonline/enthuware/exam815/LambdaExpression.java
@@ -1,4 +1,4 @@
-package com.vkeonline.enthuware;
+package com.vkeonline.enthuware.exam815;
 
 import java.util.List;
 import java.util.function.Predicate;
