+++ jdk11ocp/features/src/main/java/com/vkeonline/enthuware/exam815/PrimitiveTypes.java
@@ -1,4 +1,4 @@
-package com.vkeonline.studyguide;
+package com.vkeonline.enthuware.exam815;
 
 
 /**
