+++ jdk11ocp/features/src/main/java/com/vkeonline/enthuware/exam815/UsingConstructors.java
@@ -1,4 +1,4 @@
-package com.vkeonline.enthuware;
+package com.vkeonline.enthuware.exam815;
 
 
 class Game {
