+++ jdk11ocp/features/src/main/java/com/vkeonline/enthuware/exam816/AnnotationFeatures.java
@@ -1,4 +1,4 @@
-package com.vkeonline.enthuware;
+package com.vkeonline.enthuware.exam816;
 
 import java.lang.annotation.*;
 import java.util.HashMap;
