+++ jdk11ocp/features/src/main/java/com/vkeonline/enthuware/exam816/NIO2Features.java
@@ -1,4 +1,4 @@
-package com.vkeonline.studyguide;
+package com.vkeonline.enthuware.exam816;
 
 import java.io.File;
 import java.io.IOException;
