+++ jdk11ocp/features/src/main/java/com/vkeonline/enthuware/exam816/SecurityFeatures.java
@@ -1,4 +1,7 @@
 package com.vkeonline.enthuware.exam816;
 
+/**
+ * @author csgear
+ */
 public class SecurityFeatures {
 }
