--- jdk11ocp/features/src/main/java/com/vkeonline/studyguide/HttpClientFeature.java
@@ -1,26 +1,0 @@
-package com.vkeonline.studyguide;
-
-import java.io.IOException;
-import java.net.URI;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
-
-public class HttpClientFeature {
-    public void getUrl() {
-        try {
-            HttpClient client = HttpClient.newHttpClient();
-            HttpRequest request = HttpRequest.newBuilder()
-                    .uri(URI.create("http://webcode.me"))
-                    .build();
-
-            HttpResponse<String> response = client.send(request,
-                    HttpResponse.BodyHandlers.ofString());
-
-            System.out.println(response.body());
-        } catch (IOException | InterruptedException ignored) {
-
-        }
-
-    }
-}
