--- jdk11ocp/features/src/main/java/com/vkeonline/studyguide/JDBCFeatures.java
@@ -1,4 +1,0 @@
-package com.vkeonline.studyguide;
-
-public class JDBCFeatures {
-}
