+++ jdk11ocp/features/src/test/java/com/vkeonline/enthuware/StringFeaturesTest.java
@@ -1,8 +1,7 @@
 package com.vkeonline.enthuware;
 
+import com.vkeonline.enthuware.exam815.StringFeatures;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class StringFeaturesTest {
     static StringFeatures sf = new StringFeatures() ;
