--- jdk11ocp/features/src/test/java/com/vkeonline/studyguide/HttpClientFeatureTest.java
@@ -1,11 +1,0 @@
-package com.vkeonline.studyguide;
-
-import org.junit.jupiter.api.Test;
-
-class HttpClientFeatureTest {
-    static HttpClientFeature hcf = new HttpClientFeature() ;
-    @Test
-    void getUrl() {
-        hcf.getUrl();
-    }
-}
